+++ Encoding/ffmpegEnc.vb
@@ -70,6 +70,8 @@ Public Class ffmpegEnc
                     Return "avi"
                 Case "ProRes"
                     Return "mov"
+                Case "VP8", "VP9"
+                    Return "webm"
                 Case Else
                     Return "mkv"
             End Select
@@ -140,8 +142,8 @@ Public Class ffmpegEnc
             .Switch = "-c:v",
             .Text = "Codec",
             .AlwaysOn = True,
-            .Options = {"x264", "x265", "VP9", "Xvid", "ASP", "Theora", "ProRes", "UT Video", "Intel H.264", "Intel H.265", "NVIDIA H.264", "NVIDIA H.265"},
-            .Values = {"libx264", "libx265", "libvpx-vp9", "libxvid", "mpeg4", "libtheora", "prores", "utvideo", "h264_qsv", "hevc_qsv", "h264_nvenc", "hevc_nvenc"}}
+            .Options = {"x264", "x265", "VP8", "VP9", "Xvid", "ASP", "Theora", "ProRes", "UT Video", "Intel H.264", "Intel H.265", "NVIDIA H.264", "NVIDIA H.265"},
+            .Values = {"libx264", "libx265", "libvpx", "libvpx-vp9", "libxvid", "mpeg4", "libtheora", "prores", "utvideo", "h264_qsv", "hevc_qsv", "h264_nvenc", "hevc_nvenc"}}
 
         Property Mode As New OptionParam With {
             .Name = "Mode",
@@ -154,6 +156,10 @@ Public Class ffmpegEnc
             .Options = {"AviSynth/VapourSynth", "Intel", "DXVA2"},
             .Values = {"avs", "qsv", "dxva2"}}
 
+        Property Custom As New StringParam With {
+            .Text = "Custom",
+            .AlwaysOn = True}
+
         Overrides ReadOnly Property Items As List(Of CommandLineParam)
             Get
                 If ItemsValue Is Nothing Then
@@ -163,7 +169,7 @@ Public Class ffmpegEnc
                         New OptionParam With {.Name = "x264/x265 preset", .Text = "Preset", .Switch = "-preset", .InitValue = 5, .Options = {"Ultrafast", "Superfast", "Veryfast", "Faster", "Fast", "Medium", "Slow", "Slower", "Veryslow", "Placebo"}, .VisibleFunc = Function() Codec.OptionText.EqualsAny("x264", "x265")},
                         New OptionParam With {.Name = "x264/x265 tune", .Text = "Tune", .Switch = "-tune", .Options = {"None", "Film", "Animation", "Grain", "Stillimage", "Psnr", "Ssim", "Fastdecode", "Zerolatency"}, .VisibleFunc = Function() Codec.OptionText.EqualsAny("x264", "x265")},
                         New OptionParam With {.Switch = "-profile:v", .Text = "Profile", .VisibleFunc = Function() Codec.OptionText = "ProRes", .InitValue = 3, .IntegerValue = True, .Options = {"Proxy", "LT", "Normal", "HQ"}},
-                        New OptionParam With {.Switch = "-speed", .Text = "Speed", .AlwaysOn = True, .VisibleFunc = Function() Codec.OptionText = "VP9", .Options = {"6 - Fastest", "5 - Faster", "4 - Fast", "3 - Medium", "2 - Slow", "1 - Slower", "0 - Slowest"}, .Values = {"6", "5", "4", "3", "2", "1", "0"}, .Value = 5},
+                        New OptionParam With {.Switch = "-speed", .Text = "Speed", .AlwaysOn = True, .VisibleFunc = Function() Codec.OptionText.EqualsAny("VP8", "VP9"), .Options = {"6 - Fastest", "5 - Faster", "4 - Fast", "3 - Medium", "2 - Slow", "1 - Slower", "0 - Slowest"}, .Values = {"6", "5", "4", "3", "2", "1", "0"}, .Value = 5},
                         New OptionParam With {.Switch = "-aq-mode", .Text = "AQ Mode", .VisibleFunc = Function() Codec.OptionText = "VP9", .Options = {"Disabled", "0", "1", "2", "3"}, .Values = {"Disabled", "0", "1", "2", "3"}},
                         New OptionParam With {.Name = "h264_nvenc profile", .Switch = "-profile", .Text = "Profile", .Options = {"Baseline", "Main", "High", "High444p"}, .InitValue = 1, .VisibleFunc = Function() Codec.ValueText = "h264_nvenc"},
                         New OptionParam With {.Name = "h264_nvenc preset", .Switch = "-preset", .Text = "Preset", .Options = {"Default", "Slow", "Medium", "Fast", "HP", "HQ", "BD", "LL", "LLHQ", "LLHP", "Lossless", "Losslesshp"}, .InitValue = 2, .VisibleFunc = Function() Codec.ValueText = "h264_nvenc"},
@@ -173,14 +179,14 @@ Public Class ffmpegEnc
                         New OptionParam With {.Name = "utVideoPixFmt", .Switch = "-pix_fmt", .Text = "Pixel Format", .Options = {"YUV420P", "YUV422P", "YUV444P", "RGB24", "RGBA"}, .VisibleFunc = Function() Codec.ValueText = "utvideo"},
                         New NumParam With {.Name = "Quality", .Text = "Quality", .VisibleFunc = Function() Mode.Value = EncodingMode.Quality AndAlso Not Codec.ValueText.EqualsAny("prores", "utvideo"), .ArgsFunc = AddressOf GetQualityArgs, .Config = {0, 63}},
                         New NumParam With {.Switch = "-threads", .Text = "Decoding Threads", .Config = {0, 64}},
-                        New NumParam With {.Name = "VP9 enc threads", .Switch = "-threads", .Text = "Encoding Threads", .VisibleFunc = Function() Codec.OptionText = "VP9", .Value = 8, .DefaultValue = -1},
+                        New NumParam With {.Name = "VPX enc threads", .Switch = "-threads", .Text = "Encoding Threads", .VisibleFunc = Function() Codec.OptionText.EqualsAny("VP8", "VP9"), .Value = 8, .DefaultValue = -1},
                         New NumParam With {.Switch = "-tile-columns", .Text = "Tile Columns", .VisibleFunc = Function() Codec.OptionText = "VP9", .Value = 6, .DefaultValue = -1},
                         New NumParam With {.Switch = "-frame-parallel", .Text = "Frame Parallel", .VisibleFunc = Function() Codec.OptionText = "VP9", .Value = 1, .DefaultValue = -1},
                         New NumParam With {.Switch = "-auto-alt-ref", .Text = "Auto Alt Ref", .VisibleFunc = Function() Codec.OptionText = "VP9", .Value = 1, .DefaultValue = -1},
                         New NumParam With {.Switch = "-lag-in-frames", .Text = "Lag In Frames", .VisibleFunc = Function() Codec.OptionText = "VP9", .Value = 25, .DefaultValue = -1},
                         New BoolParam With {.Name = "h264_qsv Lookahead", .Text = "Lookahead", .Switch = "-look_ahead 1", .NoSwitch = "-look_ahead 0", .Value = False, .DefaultValue = True, .VisibleFunc = Function() Codec.ValueText = "h264_qsv"},
                         New BoolParam With {.Name = "h264_qsv VCM", .Text = "VCM", .Switch = "-vcm 1", .NoSwitch = "-vcm 0", .Value = False, .DefaultValue = True, .VisibleFunc = Function() Codec.ValueText = "h264_qsv"},
-                        New StringParam With {.Text = "Custom", .AlwaysOn = True}})
+                        Custom})
                 End If
 
                 Return ItemsValue
@@ -205,9 +211,8 @@ Public Class ffmpegEnc
             End Select
 
             If includePaths Then ret += " -i " + sourcePath.Escape
-            Dim items = From i In Me.Items Where i.GetArgs <> ""
-            If items.Count > 0 Then ret += " " + items.Select(Function(item) item.GetArgs).Join(" ")
-
+            Dim items = From i In Me.Items Where i.GetArgs <> "" AndAlso Not IsCustom(i.Switch)
+            If items.Count > 0 Then ret += " " + items.Select (Function(item) item.GetArgs).Join(" ")
             If Calc.IsARSignalingRequired Then ret += " -aspect " + Calc.GetTargetDAR.ToInvariantString.Shorten(8)
 
             Select Case Mode.Value
@@ -220,7 +225,6 @@ Public Class ffmpegEnc
             End Select
 
             If Codec.OptionText = "Xvid" Then ret += " -tag:v xvid"
-
             Dim targetPath As String
 
             If Mode.OptionText = "Two Pass" AndAlso pass = 1 Then
@@ -233,6 +237,11 @@ Public Class ffmpegEnc
             Return ret.Trim
         End Function
 
+        Function IsCustom(switch As String) As Boolean
+            If switch = "" Then Return False
+            If Custom.Value?.Contains(switch + " ") OrElse Custom.Value?.EndsWith(switch) Then Return True
+        End Function
+
         Public Overrides Function GetPackage() As Package
             Return Package.ffmpeg
         End Function
@@ -242,7 +251,7 @@ Public Class ffmpegEnc
                 Dim param = GetNumParamByName("Quality")
 
                 If param.Value <> param.DefaultValue Then
-                    If Codec.OptionText = "VP9" Then
+                    If Codec.OptionText.EqualsAny("VP8", "VP9") Then
                         Return "-crf " & param.Value & " -b:v 0"
                     ElseIf Codec.OptionText.EqualsAny("x264", "x265") Then
                         Return "-crf " & param.Value